mfdatafeed/requirements.txt
azure-core==1.34.0
azure-functions==1.23.0
azure-identity==1.23.0
azure-keyvault-secrets==4.10.0
azure-monitor-opentelemetry-exporter==1.0.0b38
azure-storage-blob==12.25.1
beautifulsoup4==4.13.4
bs4==0.0.2
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
dnspython==2.7.0
fixedint==0.1.6
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
mysql-connector-python==9.4.0
oauthlib==3.3.0
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.23.0
PyJWT==2.10.1
pymongo==4.13.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyzipper==0.3.6
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
Werkzeug==3.1.3
zipp==3.23.0
zohocrmsdk2_0==6.0.0
